arcdico/pyproject.toml
[tool.poetry]
name = "arcdico"
version = "0.1.0"
description = "Python module to interact with the ArC Instruments Digital Control Module"
authors = ["Spyros Stathopoulos <spyros@arc-instruments.co.uk>"]
readme = "README.md"
homepage = "http://www.arc-instruments.co.uk/products/arc-one/"
repository = "https://github.com/arc-instruments/arc-dico"
license = "MPL-2.0"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3 :: 3.7",
    "Programming Language :: Python :: 3 :: 3.8",
    "Programming Language :: Python :: 3 :: 3.9",
    "Programming Language :: Python :: 3 :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware",
]

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.0"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/arc-instruments/arc-dico/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
